Guard against somehow non-existant files in notmuch archive script

+++ notmuch_archive.sh
@@ -7,8 +7,10 @@ shopt -sq extglob
 while IFS= read -r -d $'\0' path; do
     file=$(basename "${path}")
     esc_file=${file//,U=[0-9]*([0-9])/}
-    mv -n "${path}" "${HOME}/.mail/fastmail/Archive/cur/${esc_file}"
-    rm -f "${path}"
+    if [[ -f "${path}" ]]; then
+        mv -n "${path}" "${HOME}/.mail/fastmail/Archive/cur/${esc_file}"
+        rm -f "${path}"
+    fi
 done < <(notmuch search --output=files --format=text0 -- folder:Inbox and not tag:inbox)
 
 notmuch new
